lib: including medicines page
Following the images on Figma, there is a fourth tab on navigation bar
destinated to Medicines reminders. This first commit adds this tab
button by moving the cards code from the main page to its own page, also
adding the header.

This commit will be squashed with all commits related to this medicines
page in the future. So far, this is only for restoring reasons.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>