Improves Remedios page
Implements several features related to the medicines manager page. Adds
a header where user can select the date and record a new medicine to
take, where it can be seen the user summary, and a list with all the
medicines in the day, where the user can click in a button to edit it.

TODO: In the Medicine class there is two variables for when the medicine
must begin to be taken and one for when the medicine must end. In this
page, this data is taken, but is not actually revelant, since it doesn't
affect anything else. One thing that can be done to improve this is to
use this information to show medicines depending on the day.

TODO: When someone takes a lot of pills in a single day, it can be
tricky to remember which ones are already taken, so one cool future
feature is to allow the user to select a checkbox in each medicine and
mark it as taken.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>